CHANGED 5 files
ADDED 4 files.
UPDATED 1 file.

Added version 5 of the algorithm, which uses a QuadTree data structure which allows less comparisons. This leads to higher FPS when running the code and the ability to include more Boids in the "world". Also added a visual representation of the QuadTree and a Play/Pause option.